Former-commit-id: 17c33e7f5e745cabee6c95dd7c0f4187b2cb2155

+++ src/main/java/org/codealpha/gmsservice/controllers/GrantController.java
@@ -567,27 +567,32 @@ public class GrantController {
             @ApiParam(name = "userId", value = "Unique identifier of logged in user") @PathVariable("userId") Long userId,
             @ApiParam(name = "X-TENANT-CODE", value = "Tenant code ") @RequestHeader("X-TENANT-CODE") String tenantCode) {
         Grant grant = grantService.getById(grantId);
-        for (GrantSpecificSection section : grantService.getGrantSections(grant)) {
-            List<GrantSpecificSectionAttribute> attribs = grantService.getAttributesBySection(section);
-            for (GrantSpecificSectionAttribute attribute : attribs) {
-                List<GrantStringAttribute> strAttribs = grantService.getStringAttributesByAttribute(attribute);
-                grantService.deleteStringAttributes(strAttribs);
+        if (grantService.checkIfGrantMovedThroughWFAtleastOnce(grantId)) {
+            grant.setDeleted(true);
+            grantService.saveGrant(grant);
+        } else {
+            for (GrantSpecificSection section : grantService.getGrantSections(grant)) {
+                List<GrantSpecificSectionAttribute> attribs = grantService.getAttributesBySection(section);
+                for (GrantSpecificSectionAttribute attribute : attribs) {
+                    List<GrantStringAttribute> strAttribs = grantService.getStringAttributesByAttribute(attribute);
+                    grantService.deleteStringAttributes(strAttribs);
+                }
+                grantService.deleteSectionAttributes(attribs);
+                grantService.deleteSection(section);
             }
-            grantService.deleteSectionAttributes(attribs);
-            grantService.deleteSection(section);
+            if (grant.getOrigGrantId() != null) {
+                Grant origGrant = grantService.getById(grant.getOrigGrantId());
+                origGrant.setAmendGrantId(null);
+                origGrant.setAmended(false);
+                grantService.saveGrant(origGrant);
+            }
+            grantService.deleteGrant(grant);
+            GranterGrantTemplate template = granterGrantTemplateService.findByTemplateId(grant.getTemplateId());
+            if (!template.isPublished()) {
+                grantService.deleteGrantTemplate(template);
+            }
         }
-        if (grant.getOrigGrantId() != null) {
-            Grant origGrant = grantService.getById(grant.getOrigGrantId());
-            origGrant.setAmendGrantId(null);
-            origGrant.setAmended(false);
-            grantService.saveGrant(origGrant);
-        }
-        grantService.deleteGrant(grant);
 
-        GranterGrantTemplate template = granterGrantTemplateService.findByTemplateId(grant.getTemplateId());
-        if (!template.isPublished()) {
-            grantService.deleteGrantTemplate(template);
-        }
     }
 
     @PostMapping("/{grantId}/section/{sectionId}/field")
@@ -1702,7 +1707,8 @@ public class GrantController {
         }
 
         // Save Snapshot
-        _saveSnapShot(grant, fromStateId, currentDateTime,
+
+        _saveSnapShot(grant, fromStateId, toStateId, grant.getNote(), currentDateTime,
                 assignmentForCurrentState.isPresent()
                         ? userService.getUserById(assignmentForCurrentState.get().getAssignments())
                         : null,
@@ -2315,7 +2321,8 @@ public class GrantController {
         }
     }
 
-    private void _saveSnapShot(Grant grant, Long fromStateId, Date movedOn, User currentUser, User previousUser) {
+    private void _saveSnapShot(Grant grant, Long fromStateId, Long toStatusId, String fromNote, Date movedOn,
+            User currentUser, User previousUser) {
 
         try {
             GrantSnapshot snapshot = new GrantSnapshot();
@@ -2334,7 +2341,10 @@ public class GrantController {
             snapshot.setName(grant.getName());
             snapshot.setRepresentative(grant.getRepresentative());
             snapshot.setStartDate(grant.getStartDate());
-            snapshot.setGrantStatusId(fromStateId);
+            snapshot.setGrantStatusId(fromStateId); // Legacy for backward compatibility
+            snapshot.setToStateId(toStatusId);
+            snapshot.setFromStateId(fromStateId);
+            snapshot.setFromNote(fromNote);
             snapshot.setStringAttributes(new ObjectMapper().writeValueAsString(grant.getGrantDetails()));
             grantSnapshotService.saveGrantSnapshot(snapshot);
         } catch (JsonProcessingException e) {
@@ -2978,11 +2988,30 @@ public class GrantController {
     public List<GrantHistory> getGrantHistory(@PathVariable("grantId") Long grantId,
             @PathVariable("userId") Long userId, @RequestHeader("X-TENANT-CODE") String tenantCode) {
 
-        List<GrantHistory> history = grantService.getGrantHistory(grantId);
-        for (GrantHistory historyEntry : history) {
-            historyEntry.setNoteAddedByUser(userService.getUserByEmailAndOrg(historyEntry.getNoteAddedBy(),
-                    historyEntry.getGrantorOrganization()));
+        List<GrantHistory> history = new ArrayList();
+        List<GrantSnapshot> grantSnapshotHistory = grantSnapshotService.getGrantSnapshotForGrant(grantId);
+        if (grantSnapshotHistory == null
+                || (grantSnapshotHistory != null && grantSnapshotHistory.get(0).getFromStateId() == null)) {
+            history = grantService.getGrantHistory(grantId);
+            for (GrantHistory historyEntry : history) {
+                historyEntry.setNoteAddedByUser(userService.getUserByEmailAndOrg(historyEntry.getNoteAddedBy(),
+                        historyEntry.getGrantorOrganization()));
+            }
+        } else {
+            for (GrantSnapshot snapShot : grantSnapshotHistory) {
+                GrantHistory hist = new GrantHistory();
+                hist.setName(snapShot.getName());
+                hist.setId(snapShot.getGrantId());
+                hist.setNote(snapShot.getFromNote());
+                hist.setNoteAdded(snapShot.getMovedOn());
+                User assignedBy = userService.getUserById(snapShot.getAssignedBy());
+                hist.setNoteAddedBy(assignedBy.getFirstName().concat(" ").concat(assignedBy.getLastName()));
+                hist.setNoteAddedByUser(assignedBy);
+                hist.setGrantStatus(workflowStatusService.findById(snapShot.getFromStateId()));
+                history.add(hist);
+            }
         }
+
         return history;
     }
